requirements/base.txt
langchain
# langchain-openai 
langchainhub
langchain-community
# tavily-python
langchain-groq
langchain-google-genai
langchain-core
# langchain-google-vertexai
pypdf
docx2txt
langchain-text-splitters
chromadb
fastapi 
uvicorn 
python-multipart
pydantic[all]
duckduckgo-search
numexpr
python-dotenv
unstructured
# markdown
# langchain_experimental
# unstructured[pdf]
# unstructured[doc_type]
# unstructured[pptx]
sentence-transformers
## Auth utilities 
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy 
psycopg2